Doctype support
I wasn't going to do this before because it still wouldn't handle all
well-formed XML files because you could have a conditional section.

But then I realized conditional sections are only allowed in external
doctypes. So this should mean that, excepting bugs, we're able to
parse all well-formed documents encoded in UTF-8. That's worth a few
lines of code.